Fix harfbuzz font lifecycle in CommonSalLayout
The harfbuzz font is attached to the system font face and
therefore inherits its lifecycle. This means it can be used in
multiple CommonSalLayout objects, so the user data parameter of
hb_face_create_for_tables can't be the layout, but must be the
font.

This moves the special Qt5Font handling into it's own function,
so accessing the switching parameter mbUseQt5 is not needed.

Regression from commit 97f61dc6e2cad79e94dd9faa83739f0d9d99408a.

Change-Id: Ic34cc5b60e401562c73b239a58176a59fe4bf9be
Reviewed-on: https://gerrit.libreoffice.org/44398
Reviewed-by: Khaled Hosny <khaledhosny@eglug.org>
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>