tdf#113428: Fix glyph positions in PDF export
Not sure what is going on, but PDF export is the only place we call
GetNextGlyphs() with nLen > 1 and it seems I broke something here in
af4b5dfdc7e349fd8602b798a65d3d2baed3077c. Instead of trying to figure out
what the old code was trying to do, lets just calculate glyph positions
for the extra glyphs as GetNextGlyphs() does for the first one, making
PDF export closer to how the rendering on screen works.

Change-Id: I3f516e954ccfbe38abcf6985efe3bb1f998e967f
Reviewed-on: https://gerrit.libreoffice.org/44514
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Khaled Hosny <khaledhosny@eglug.org>