TSCP: bail out quickly in paragraph classification check on save
Before saving a document we need to check that the document level
classification is at the higher or same level as the individual
paragraph level classification is. The trouble here is tha this
can be quite a costly operation, so bail out quickly if we don't
detect that no paragraph classification has been applied to the
content by checking if we have the RDF metadata present or not.

Change-Id: Iaef29de7e138937a5c1881cf34c9075fb37ba87d
Reviewed-on: https://gerrit.libreoffice.org/44573
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>