TSCP: resolve identifier to name, abbr. name is for category only
In paragraph classification we store only the identifier and not
the name, but for the ClassificationDialog we use the name as an
identifier, so we need to resovle the identifier to the name when
reading in the input classification results. It might make sense
to switch to identifiers in the dialog also as they are language
independent.

Abbreviated name is only for the classification category and has
no effect for other classefication elementslike markers and IP
parts, so always fill them with the full name directly.

Change-Id: I23bd0e8e92b16807a5b4d3162e503b799aa40718
Reviewed-on: https://gerrit.libreoffice.org/44845
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>