gbuild-to-ide cleaned up double scripts.
Moved gbuild-to-ideNS to gbuild-to-ide (NS was the original)
Removed gbuild-to-ideNS from Makefile.in

the original gbuild-to-ide was split off to gbuild-to-ideNS, in order
not to disturb the production environment while playing with new ide
generation.

Work on a new IDE generator have been stopped for a half year, so time
to clean up.

The work on a new Xcode generator continues offline (for now), and are
part of the iOS development.

Change-Id: Ia67355df21c6aec592d792ffa7c904a568eff961