pixman: drop cached glyphs above size watermark
The glyphs cache only tracked the number of glyphs
in the cache, which isn't very useful when the
glyphs vary in size from a few bytes to dozens of
megabytes. In terms of memory footprint, we are
interested not just in the number of cached glyphs,
but also in the total memory usage.

We now track the pixel-size of each glyph and drop
as many as necessary to fall below the set threshold
which is currently 4 MBs.

There was no performance loss observed after this patch
on a diverse set of documents of all types (if anything,
a very modest performance win was hinted in some cases),
with significant memory reduction in some cases where
dozens of megabytes were previously left stale in the cache.

Change-Id: I377a4c1e59611fca6bf253dd9fa90f75b5295d4b
Reviewed-on: https://gerrit.libreoffice.org/45378
Reviewed-by: Ashod Nakashian <ashnakash@gmail.com>
Tested-by: Ashod Nakashian <ashnakash@gmail.com>
(cherry picked from commit e8b5a5e1f0a50d91afa4f1707f18641ab94bcafd)
(cherry picked from commit ded39dc17a34127a872b11abddc32bbbb5aae047)
Reviewed-on: https://gerrit.libreoffice.org/46987
Tested-by: Jenkins <ci@libreoffice.org>