undo "ofz#3577 force refresh"
oss-fuzz shows no reduction of timeouts. Seeing as this
is the root of importing ole docproperties none of the lines
from SfxOleStringHelper:: up can be in any timeouts, so the
original stacktrace isn't in any way helpful and it must
be the generic ReadUInt16 downwards part of the stack.

In other words the timeout bugs in oss-fuzz are useless if they
aren't reproducible

This reverts commit 8b5fb2a71b8dc027a397a5cb4e586ece2ab3a785.

Change-Id: I3fb8d2ded8502278098bd223898e3e90f6e4a378
Reviewed-on: https://gerrit.libreoffice.org/45352
Reviewed-by: Caolán McNamara <caolanm@redhat.com>
Tested-by: Caolán McNamara <caolanm@redhat.com>