ODF export: expose font style and weight of embedded fonts
We don't need this info ourselves, but when converting the flat ODF
output to other formats, this information is hard to find out, so state
them explicitly. Names correspond to existing CSS ones, where it's
necessary to state these explicitly. The EPUB export will make use of
these attributes.

(It is unclear why the test fails on the windows_msc_dbgutil_32 slave,
it works for me locally.)

Change-Id: I05030b6a14aca842e5a740786f3743e9ae838293
Reviewed-on: https://gerrit.libreoffice.org/45384
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Miklos Vajna <vmiklos@collabora.co.uk>