Remove arch-pinning for gpgme on Windows again
With e245d122c001d177fc719d55f88b55b04f7aa422 the dependency on gcc
is gone, so we should be able to build on all cygwin flavours again
for all Windows arch targets.

Well almost - we still have windres on cygwin32 only able to produce
pe-i386 targets, so let's exclude the case cyg32 and WIN64 for the
moment.

Change-Id: Ibe7f967585995586985177d4c36cc33d9850d9de
Reviewed-on: https://gerrit.libreoffice.org/45422
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Thorsten Behrens <Thorsten.Behrens@CIB.de>