xmloff: turn SvXMLImportContext into a no-op base class
Fix the non-obvious and potentially dangerous recursion that is
implicit in xmloff's context code.

SvXMLImportContext::CreateChildContext() now always creates a
SvXMLImportContext, does not delegate to SvXMLImport.

Rename CreateContext to CreateDocumentContext, to make its purpose of
creating only the top-level context more obvious.

With the previous changes to CreateContext overrides in sw,
reportdesign and dbaccess, this hopefully shouldn't break anything.

Change-Id: I9e147bf6297bbac9e8470454881da73f6e39db0a