lokdialog: Show/Hide floating window on StateChangedType::Visible
In the earlier approach, other state change events like SetText are
emitted which ends up in "else if !IsVisible()" block and tries to close
the floating window before it's even registered.

Interestingly, this is not the case every time. Anyhow, better safe than
sorry. Guard the floating window 'created', 'close' callbacks in a
StateChangeType::Visible event.

Change-Id: Ib997e11fd4a0f50b8911e9891918112e8ff8ef85