Resolves: tdf#114406 treat % as the operator that it is
Regression from

    commit 419856e962390dd09b3e8862a0a1f184882b5c92
    Date:   Tue Sep 20 21:39:10 2016 +0200

	sc-perf: tdf#79023 for ODFF do not call SvNumberFormatter to determine numeric

for ODFF, and for OOXML

    commit 7fe6d5be7f9f3f20be9fcfc5783562c0cc6cab98
    Date:   Tue Sep 20 22:37:59 2016 +0200

	sc-perf: tdf#79023 do not call SvNumberFormatter also for numbers in OOXML

Numbers followed by a percent operator like 100% were always
(wrongly) treated as one entity (by '%' having ScCharFlags::Value)
and incidentally handled by the number formatter's parser, which
so far "worked" in the sense that the correct constant number was
produced (i.e. 1 here), but the expression detail was lost.

The commits above for performance reasons when reading ODF or
OOXML files don't let a symbol pass through the number formatter's
parser anymore and the "100%" does not represent a strict number,
so lead to #NAME? error at the end.

Set the proper flags in the compiler's character table to treat
'%' as operator in all circumstances.

Change-Id: I266beb74a313c779370e5fac42f45d0fb5cdba0c