show rest of the menu bar in a pop-up if the are is too small
In case when the window is to small to show the whole menu then
the use couldn't access the rest of the items in the menu. This
change adds a button in the end (using the ">>" marker as in
toolbar) which on click adds a pop-up that shows the rest of the
menu items.

Change-Id: I9218dba504464bdd44d61ebb383f7674f3df760f
Reviewed-on: https://gerrit.libreoffice.org/46711
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>